+++ Expence.h
@@ -9,20 +9,20 @@ class Expence
 {
     int id, userId, date;
     string item;
-    double money;
+    float money;
 
 public:
     void setId(int newId);
     void setUserId(int newUserId);
     void setDate(int newDate);
     void setItem(string newItem);
-    void setMoney(double newMoney);
+    void setMoney(float newMoney);
 
 
     int downloadId();
     int downloadUserId();
     int downloadDate();
     string downloadItem();
-    double downloadMoney();
+    float downloadMoney();
 };
 #endif
